requirements.txt
python-dotenv
tweepy
pocket
